service/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.24.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-dotenv==1.0.1
google-cloud-storage==2.17.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-api-core==2.20.0
pypdf==4.2.0
Pillow==10.4.0
openpyxl==3.1.5
